requirements.txt
graphene-django
django-graphql-jwt
